MediaCatalog/requirements.txt
APScheduler==3.6.3
asgiref==3.2.3
aws-requests-auth==0.4.2
blist==1.3.6
boto3==1.10.44
botocore==1.13.44
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
configparser==4.0.2
coreapi==2.3.3
coreschema==0.0.4
croniter==0.3.30
defusedxml==0.6.0
Django==2.2.9
django-elasticsearch-dsl-drf==0.20.3
django-filter==2.2.0
django-guardian==2.1.0
django-haystack==2.8.1
django-nine==0.2.2
django-oauth-toolkit==1.2.0
django-rest-framework==0.1.0
django-six==1.0.4
djangorestframework==3.10.3
docopt==0.6.2
docutils==0.15.2
drf-haystack==1.8.6
elastalert==0.2.1
elasticsearch==6.4.0
elasticsearch-dsl==7.1.0
envparse==0.2.0
exotel==0.1.5
future==0.18.2
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
jira==1.0.14
jmespath==0.9.4
jsonschema==2.6.0
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
oauthlib==3.1.0
pbr==5.4.4
psycopg2==2.8.4
pycparser==2.19
Pygments==2.5.2
PyJWT==1.7.1
PySocks==1.7.1
PyStaticConfiguration==0.10.4
python-dateutil==2.6.1
python-magic==0.4.15
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
stomp.py==4.1.22
texttable==1.6.2
thehive4py==1.6.0
twilio==6.0.0
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.25.7
